refactor: changed folder name aboutComponents to about
