+++ src/features/about/AboutHero.tsx
